requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
openai==0.28.1
json-repair==0.41.1
python-multipart==0.0.9
requests==2.31.0 